rename to pyirf
