+++ MANIFEST.in
@@ -1,6 +1,6 @@
 include README.rst
 
-recursive-include pyrf/resources
-recursive-include pyrf/scripts/*.py
+recursive-include pyirf/resources
+recursive-include pyirf/scripts/*.py
 
 global-exclude *.pyc
