+++ docs/index.rst
@@ -1,20 +1,20 @@
-.. pyrf documentation master file, created by
+.. pyirf documentation master file, created by
    sphinx-quickstart on Sat Apr 25 16:39:28 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to pyrf's documentation!
+Welcome to pyirf's documentation!
 ================================
 
-pyrf can:
+pyirf can:
 
   * find the best cutoff in gammaness/score, to discriminate between signal
     and background, as well as the angular cut to obtain the best sensitivity
     for a given amount of observation time and a given template for the
-    source of interest (:ref:`pyrf`)
+    source of interest (:ref:`pyirf`)
   * compute the instrument response functions, effective area,
-    point spread function and energy resolution (:ref:`pyrf`)
-  * estimate the sensitivity (:ref:`pyrf`)
+    point spread function and energy resolution (:ref:`pyirf`)
+  * estimate the sensitivity (:ref:`pyirf`)
 
 
 .. toctree::
