+++ pyirf/perf/cut_optimisation.py
