+++ pyirf/resources/performance.yml
