+++ setup.cfg
@@ -1,12 +1,12 @@
 [metadata]
-name = pyrf
+name = pyirf
 description = Python IACT IRF builder,
-url = https://github.com/cta-observatory/pyrf
+url = https://github.com/cta-observatory/pyirf
 author = Julien Lefaucheur, Michele Peresano, Thomas Vuillaume
 author_email = thomas.vuillaume@lapp.in2p3.fr
 license = MIT
 long_description = file: README.rst
-github_project = cta-observatory/pyrf
+github_project = cta-observatory/pyirf
 
 [options]
 python_requires = >=3.6
